gammatone init: comments from review
